pyproject.toml
[project]
name = "doceval"
version = "0.1.0"
description = "Evaluate LLMs on legal document generation and editing — quality, speed, cost."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=0.3.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "choix>=0.3.5",
    "numpy>=1.26.0",
]

[project.scripts]
doceval = "doceval.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["doceval"]

[tool.pytest.ini_options]
testpaths = ["tests"]
